Enable noVNC to become Browserifiable
This commit restructures noVNC, splitting it into the core directory
and the app directory, with the former containing core noVNC parts,
and the latter containing parts specific to the application.

Additionally, it introduces a new utility for transforming the noVNC
javascript files into a browserifiable application.  Module names and
requirements are listed at the top of each file, and the script
`use_require.js` may be used to convert them into a form that uses
`module.exports` and `require`.  If you pass the '-b' flag to
the script, it will aslo run browserify, and copy over sufficient
resources to use 'vnc.html'.